Extract render_single_pane to pane_render.rs (refs #68)
Moves the 540-line render_single_pane() method out of main.rs into a
dedicated pane_render.rs module. Renders each pane's tab bar (surface
tabs, click/middle-click/right-click/drag/rename/close, trailing "+")
and delegates terminal content rendering to the GPU or software path.

Mirrors wezterm-gui's termwindow/render/pane.rs — the per-pane render
layer sitting between the paint entry and per-line cell emission.

Single biggest extraction in the amux-app phase 4 decomposition:
main.rs drops from 1,383 to 843 LoC.

No behavior change — pure mechanical refactor.
